requirements.txt
python_docx==1.1.0
selenium==3.5.0
